requirements.txt
# FastAPI and server
fastapi==0.128.8
uvicorn==0.39.0
starlette==0.49.3
python-multipart==0.0.20

# Database
SQLAlchemy==2.0.46
psycopg2-binary==2.9.11
alembic==1.16.5

# Config and validation
pydantic==2.12.5
pydantic-settings==2.11.0
pydantic_core==2.41.5
python-dotenv==1.2.1

# Authentication
python-jose[cryptography]==3.3.0
requests==2.32.3

# PDF export
fpdf2==2.8.4
pillow==11.3.0

# AI/LLM (LangChain)
langchain==0.3.27
langchain-openai==0.3.35
langgraph==0.6.11
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
pinecone==5.4.0
python-docx==1.1.2
pypdf==5.1.0
PyYAML==6.0.3
async-timeout==4.0.3
httpx==0.28.1
httpcore==1.0.9
orjson==3.11.5
ormsgpack==1.11.0
tenacity==9.1.2
jsonpatch==1.33
jsonpointer==3.0.0
requests-toolbelt==1.0.0
uuid-utils==0.14.0
xxhash==3.6.0
zstandard==0.25.0
packaging==25.0

# Google Sheets
gspread==6.1.4
google-auth==2.38.0

# Dependencies (transitive)
annotated-types==0.7.0
anyio==4.12.1
click==8.1.8
greenlet==3.2.4
h11==0.16.0
idna==3.11
Mako==1.3.10
MarkupSafe==3.0.3
tomli==2.4.0
typing-inspection==0.4.2
typing_extensions==4.15.0
fonttools==4.60.2
annotated-doc==0.0.4
defusedxml==0.7.1
exceptiongroup==1.3.1
